Add files via upload
Summary: This research paper presents a comprehensive study on the application of ML algorithms for the detection of diabetes in patients. We explore various data sources, feature selection methods, and classification algorithms to build robust predictive models. The performance of these models is evaluated using real-world healthcare datasets, demonstrating the potential for ML-based diabetes detection in clinical practice. We obtained a performance improvement of 0.03% and 0.06% using CNN and CNN-LSTM architecture respectively compared to our earlier work without using SVM. The classification system proposed can help the clinicians to diagnose diabetes using ECG signals with a very high accuracy of 95.7%.